requirements.txt
elasticsearch_dsl
matplotlib
numpy
elasticsearch
pillow
python-resize-image
